+++ tests/test_webhook.py
@@ -1,37 +1,34 @@
-from typing import TYPE_CHECKING, Dict
+import typing as t
 
 import aiohttp
+import mock
 import pytest
-from discord import Client
 
-from topgg import WebhookManager
-
-if TYPE_CHECKING:
-    from topgg.types import BotVoteData, ServerVoteData
+from topgg import WebhookManager, WebhookType
+from topgg.errors import TopGGException
 
 auth = "youshallnotpass"
 
 
 @pytest.fixture
 def webhook_manager() -> WebhookManager:
-    return WebhookManager(Client()).dbl_webhook("/dbl", auth).dsl_webhook("/dsl", auth)
+    return (
+        WebhookManager()
+        .endpoint(WebhookType.BOT)
+        .auth(auth)
+        .route("/dbl")
+        .callback(print)
+        .add_to_manager()
+        .endpoint(WebhookType.GUILD)
+        .auth(auth)
+        .route("/dsl")
+        .callback(print)
+        .add_to_manager()
+    )
 
 
 def test_WebhookManager_routes(webhook_manager: WebhookManager) -> None:
-    assert len(webhook_manager._webhooks) == 2
-
-
-@pytest.mark.asyncio
-async def test_WebhookManager_run_method(webhook_manager: WebhookManager) -> None:
-    task = webhook_manager.run(5000)
-
-    try:
-        if not task.done():
-            assert await task is None
-
-        assert hasattr(webhook_manager, "_webserver")
-    finally:
-        await webhook_manager.close()
+    assert len(webhook_manager.app.router.routes()) == 2
 
 
 @pytest.mark.asyncio
@@ -40,21 +37,9 @@ async def test_WebhookManager_run_method(webhook_manager: WebhookManager) -> Non
     [({"authorization": auth}, 200, True), ({}, 401, False)],
 )
 async def test_WebhookManager_validates_auth(
-    webhook_manager: WebhookManager, headers: Dict[str, str], result: int, state: bool
+    webhook_manager: WebhookManager, headers: t.Dict[str, str], result: int, state: bool
 ) -> None:
-    await webhook_manager.run(5000)
-
-    dbl_state = dsl_state = False
-
-    @webhook_manager.bot.event
-    async def on_dbl_vote(data: "BotVoteData") -> None:
-        nonlocal dbl_state
-        dbl_state = True
-
-    @webhook_manager.bot.event
-    async def on_dsl_vote(data: "ServerVoteData") -> None:
-        nonlocal dsl_state
-        dsl_state = True
+    await webhook_manager.start(5000)
 
     try:
         for path in ("dbl", "dsl"):
@@ -62,7 +47,22 @@ async def test_WebhookManager_validates_auth(
                 "POST", f"http://localhost:5000/{path}", headers=headers, json={}
             ) as r:
                 assert r.status == result
-
-            assert locals()[f"{path}_state"] is state
     finally:
         await webhook_manager.close()
+        assert webhook_manager._is_closed
+
+
+def test_WebhookEndpoint_callback_unset(webhook_manager: WebhookManager):
+    with pytest.raises(
+        TopGGException,
+        match=r"callback\ was\ unset,\ please\ set\ it\ using\ the\ callback\(\)\ method\.",
+    ):
+        webhook_manager.endpoint(WebhookType.BOT).add_to_manager()
+
+
+def test_WebhookEndpoint_route_unset(webhook_manager: WebhookManager):
+    with pytest.raises(
+        TopGGException,
+        match=r"route\ was\ unset,\ please\ set\ it\ using\ the\ route\(\)\ method\.",
+    ):
+        webhook_manager.endpoint(WebhookType.BOT).callback(mock.Mock()).add_to_manager()