+++ topgg/client.py
@@ -1,175 +1,86 @@
 # -*- coding: utf-8 -*-
 
-"""
-The MIT License (MIT)
+# The MIT License (MIT)
 
-Copyright (c) 2021 Assanali Mukhanov
+# Copyright (c) 2021 Assanali Mukhanov
 
-Permission is hereby granted, free of charge, to any person obtaining a
-copy of this software and associated documentation files (the "Software"),
-to deal in the Software without restriction, including without limitation
-the rights to use, copy, modify, merge, publish, distribute, sublicense,
-and/or sell copies of the Software, and to permit persons to whom the
-Software is furnished to do so, subject to the following conditions:
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
 
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
 
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-DEALINGS IN THE SOFTWARE.
-"""
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+# OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+# DEALINGS IN THE SOFTWARE.
 
-import asyncio
-import logging
-import sys
-import traceback
-from asyncio.tasks import Task
-from contextlib import suppress
-from typing import Any, Dict, List, Optional, Union, cast
+__all__ = ["DBLClient"]
 
-import discord
-from discord.ext.commands.bot import BotBase
+import typing as t
 
-# if sys.version_info >= (3, 8):
-#     from typing import TypedDict
-# else:
-#     from typing_extensions import TypedDict
+import aiohttp
+
 from . import errors, types
+from .autopost import AutoPoster
+from .data import DataContainerMixin
 from .http import HTTPClient
 
-log = logging.getLogger(__name__)
 
-
-class DBLClient:
+class DBLClient(DataContainerMixin):
     """Represents a client connection that connects to Top.gg.
     This class is used to interact with the Top.gg API.
 
-    .. _event loop: https://docs.python.org/3/library/asyncio-eventloops.html
     .. _aiohttp session: https://aiohttp.readthedocs.io/en/stable/client_reference.html#client-session
 
     Parameters
     ----------
-    bot: discord.Client
-        An instance of a discord.py Client object.
     token: str
         Your bot's Top.gg API Token.
-    autopost: bool
-        Whether to automatically post bot's guild count every 30 minutes.
-        This will dispatch :meth:`on_autopost_success` (or :meth:`on_autopost_error` in case of an error).
-    post_shard_count: bool
-        Whether to post the shard count on autopost.
-        Defaults to False.
-    autopost_interval: Optional[int]
-        Interval used by autopost to post server count automatically, measured in seconds. Defaults to 1800 (30
-        minutes) if autopost is True, otherwise None.
-    **session: :class:`aiohttp.ClientSession`
+    default_bot_id: Optional[int]
+        The default bot_id. You can override this by passing it when calling a method.
+    session: :class:`aiohttp.ClientSession`
         An `aiohttp session`_ to use for requests to the API.
     """
 
-    bot: discord.Client
-    bot_id: int
-    loop: asyncio.AbstractEventLoop
-    autopost: bool
-    post_shard_count: bool
-    _is_closed: bool
+    __slots__ = ("http", "default_bot_id", "_token", "_is_closed")
     http: HTTPClient
-    autopost_task: Task
-    autopost_interval: Optional[float]
 
     def __init__(
         self,
-        bot: discord.Client,
         token: str,
-        autopost: bool = False,
-        post_shard_count: bool = False,
-        autopost_interval: Optional[float] = None,
-        **kwargs: Any,
+        *,
+        default_bot_id: t.Optional[int] = None,
+        session: t.Optional[aiohttp.ClientSession] = None,
     ) -> None:
-        self.bot = bot
-        self.loop = bot.loop
-        self.autopost = autopost
-        self.post_shard_count = post_shard_count
-        self.autopost_interval = autopost_interval
-        self.http = HTTPClient(token, loop=self.loop, session=kwargs.get("session"))
+        super().__init__()
+        self._token = token
+        self.default_bot_id = default_bot_id
         self._is_closed = False
-
-        if not isinstance(self.autopost, bool):
-            raise errors.ClientException("autopost must be of type bool")
-        if not isinstance(self.post_shard_count, bool):
-            raise errors.ClientException("post_shard_count must be of type bool")
-        if self.autopost:
-            if self.autopost_interval is None:
-                self.autopost_interval = 1800
-            if not isinstance(self.autopost_interval, int):
-                raise errors.ClientException("autopost_interval must be of type int")
-            if self.autopost_interval < 900:
-                raise errors.ClientException(
-                    "autopost_interval must be greater than or equal to 900 seconds (15 minutes)"
-                )
-
-            if not hasattr(self.bot, "on_autopost_error"):
-                self.bot.on_autopost_error = self.on_autopost_error
-
-            self.autopost_task = self.loop.create_task(self._auto_post())
-        else:
-            if self.post_shard_count:
-                raise errors.ClientException(
-                    "autopost must be activated if post_shard_count is passed"
-                )
-            if self.autopost_interval:
-                raise errors.ClientException(
-                    "autopost must be activated if autopost_interval is passed"
-                )
-
-    async def on_autopost_error(self, exception: Exception) -> None:
-        # only print if there's no external autopost_error listeners.
-        if isinstance(self.bot, BotBase) and self.bot.extra_events.get(
-            "on_autopost_error"
-        ):
-            return
-
-        print("Ignoring exception in auto post loop:", file=sys.stderr)
-        traceback.print_exception(
-            type(exception), exception, exception.__traceback__, file=sys.stderr
-        )
-
-    async def _ensure_bot_user(self) -> None:
-        await self.bot.wait_until_ready()
-        if not hasattr(self, "bot_id") or self.bot_id is None:
-            self.bot_id = self.bot.user.id
-
-    async def _auto_post(self) -> None:
-        await self._ensure_bot_user()
-        while not self.bot.is_closed():
-            try:
-                log.debug(f"Attempting to post server count ({self.guild_count})")
-                await self.post_guild_count(
-                    shard_count=self.bot.shard_count if self.post_shard_count else None
-                )
-                event_name = "autopost_success"
-                self.bot.dispatch(event_name)
-            except Exception as e:
-                event_name = "autopost_error"
-                self.bot.dispatch(event_name, e)
-
-                if isinstance(e, errors.Unauthorized):
-                    raise
-
-            await asyncio.sleep(cast(float, self.autopost_interval))
+        if session is not None:
+            self.http = HTTPClient(token, session=session)
 
     @property
     def is_closed(self) -> bool:
         return self._is_closed
 
-    @property
-    def guild_count(self) -> int:
-        """Gets the guild count from the provided Client object."""
-        return len(self.bot.guilds)
+    async def _ensure_session(self) -> None:
+        if not hasattr(self, "http"):
+            self.http = HTTPClient(self._token, session=None)
+
+    def _validate_and_get_bot_id(self, bot_id: t.Optional[int]) -> int:
+        bot_id = bot_id or self.default_bot_id
+        if bot_id is None:
+            raise TypeError("bot_id or default_bot_id is unset.")
+
+        return bot_id
 
     async def get_weekend_status(self) -> bool:
         """This function is a coroutine.
@@ -181,14 +92,31 @@ class DBLClient:
         weekend status: bool
             The boolean value of weekend status.
         """
+        await self._ensure_session()
         data = await self.http.get_weekend_status()
         return data["is_weekend"]
 
+    @t.overload
+    async def post_guild_count(self, stats: types.StatsWrapper) -> None:
+        ...
+
+    @t.overload
     async def post_guild_count(
         self,
-        guild_count: Optional[Union[int, List[int]]] = None,
-        shard_count: Optional[int] = None,
-        shard_id: Optional[int] = None,
+        *,
+        guild_count: t.Union[int, t.List[int]],
+        shard_count: t.Optional[int] = None,
+        shard_id: t.Optional[int] = None,
+    ) -> None:
+        ...
+
+    async def post_guild_count(
+        self,
+        stats: t.Any = None,
+        *,
+        guild_count: t.Any = None,
+        shard_count: t.Any = None,
+        shard_id: t.Any = None,
     ) -> None:
         """This function is a coroutine.
 
@@ -198,6 +126,8 @@ class DBLClient:
 
         Parameters
         ----------
+        stats: :obj:`~.types.StatsWrapper`
+            An instance of StatsWrapper containing guild_count, shard_count, and shard_id.
         guild_count: Optional[Union[int, List[int]]]
             Number of guilds the bot is in. Applies the number to a shard instead if shards are specified.
             If not specified, length of provided client's property `.guilds` will be posted.
@@ -206,12 +136,18 @@ class DBLClient:
         shard_id: Optional[int]
             The index of the current shard. Top.gg uses `0 based indexing`_ for shards.
         """
-        await self._ensure_bot_user()
-        if guild_count is None:
-            guild_count = self.guild_count
+        if stats:
+            guild_count = stats.guild_count
+            shard_count = stats.shard_count
+            shard_id = stats.shard_id
+        elif guild_count is None:
+            raise TypeError("stats or guild_count must be provided.")
+        await self._ensure_session()
         await self.http.post_guild_count(guild_count, shard_count, shard_id)
 
-    async def get_guild_count(self, bot_id: Optional[int] = None) -> types.BotStatsData:
+    async def get_guild_count(
+        self, bot_id: t.Optional[int] = None
+    ) -> types.BotStatsData:
         """This function is a coroutine.
 
         Gets a bot's guild count and shard info from Top.gg.
@@ -223,16 +159,15 @@ class DBLClient:
 
         Returns
         -------
-        stats: :ref:`BotStatsData`
+        stats: :obj:`~.types.BotStatsData`
             The guild count and shards of a bot on Top.gg.
         """
-        await self._ensure_bot_user()
-        if bot_id is None:
-            bot_id = self.bot_id
+        bot_id = self._validate_and_get_bot_id(bot_id)
+        await self._ensure_session()
         response = await self.http.get_guild_count(bot_id)
         return types.BotStatsData(**response)
 
-    async def get_bot_votes(self) -> List[types.BriefUserData]:
+    async def get_bot_votes(self) -> t.List[types.BriefUserData]:
         """This function is a coroutine.
 
         Gets information about last 1000 votes for your bot on Top.gg.
@@ -242,14 +177,18 @@ class DBLClient:
 
         Returns
         -------
-        users: List[:ref:`BriefUserData`]
+        users: List[:obj:`~.types.BriefUserData`]
             Users who voted for your bot.
         """
-        await self._ensure_bot_user()
-        response = await self.http.get_bot_votes(self.bot_id)
+        if not self.default_bot_id:
+            raise errors.ClientException(
+                "you must set default_bot_id when constructing the client."
+            )
+        await self._ensure_session()
+        response = await self.http.get_bot_votes(self.default_bot_id)
         return [types.BriefUserData(**user) for user in response]
 
-    async def get_bot_info(self, bot_id: Optional[int] = None) -> types.BotData:
+    async def get_bot_info(self, bot_id: t.Optional[int] = None) -> types.BotData:
         """This function is a coroutine.
 
         Gets information about a bot from Top.gg.
@@ -261,13 +200,12 @@ class DBLClient:
 
         Returns
         -------
-        bot info: :ref:`BotData`
+        bot info: :obj:`~.types.BotData`
             Information on the bot you looked up. Returned data can be found
             `here <https://docs.top.gg/api/bot/#bot-structure>`_.
         """
-        await self._ensure_bot_user()
-        if bot_id is None:
-            bot_id = self.bot_id
+        bot_id = self._validate_and_get_bot_id(bot_id)
+        await self._ensure_session()
         response = await self.http.get_bot_info(bot_id)
         return types.BotData(**response)
 
@@ -275,10 +213,10 @@ class DBLClient:
         self,
         limit: int = 50,
         offset: int = 0,
-        sort: Optional[str] = None,
-        search: Optional[Dict[str, Any]] = None,
-        fields: Optional[List[str]] = None,
-    ) -> types.DataDict[str, Any]:
+        sort: t.Optional[str] = None,
+        search: t.Optional[t.Dict[str, t.Any]] = None,
+        fields: t.Optional[t.List[str]] = None,
+    ) -> types.DataDict[str, t.Any]:
         """This function is a coroutine.
 
         Gets information about listed bots on Top.gg.
@@ -298,12 +236,13 @@ class DBLClient:
 
         Returns
         -------
-        bots: :ref:`DataDict`
+        bots: :obj:`~.types.DataDict`
             Info on bots that match the search query on Top.gg.
         """
         sort = sort or ""
         search = search or {}
         fields = fields or []
+        await self._ensure_session()
         response = await self.http.get_bots(limit, offset, sort, search, fields)
         response["results"] = [
             types.BotData(**bot_data) for bot_data in response["results"]
@@ -322,9 +261,10 @@ class DBLClient:
 
         Returns
         -------
-        user data: :ref:`UserData`
+        user data: :obj:`~.types.UserData`
             Information about a Top.gg user.
         """
+        await self._ensure_session()
         response = await self.http.get_user_info(user_id)
         return types.UserData(**response)
 
@@ -343,19 +283,23 @@ class DBLClient:
         vote status: bool
             Info about the user's vote.
         """
-        await self._ensure_bot_user()
-        data = await self.http.get_user_vote(self.bot_id, user_id)
+        if not self.default_bot_id:
+            raise errors.ClientException(
+                "you must set default_bot_id when constructing the client."
+            )
+
+        await self._ensure_session()
+        data = await self.http.get_user_vote(self.default_bot_id, user_id)
         return bool(data["voted"])
 
-    async def generate_widget(self, options: types.WidgetOptions) -> str:
-        """This function is a coroutine.
-
-        Generates a Top.gg widget from the provided :ref:`WidgetOptions` object.
+    def generate_widget(self, *, options: types.WidgetOptions) -> str:
+        """
+        Generates a Top.gg widget from the provided :obj:`~.types.WidgetOptions` object.
 
         Parameters
         ----------
-        options: :ref:`WidgetOptions`
-            A :ref:`WidgetOptions` object containing widget parameters.
+        options: :obj:`~.types.WidgetOptions`
+            A :obj:`~.types.WidgetOptions` object containing widget parameters.
 
         Returns
         -------
@@ -366,11 +310,10 @@ class DBLClient:
             raise errors.ClientException(
                 "options argument passed to generate_widget must be of type WidgetOptions"
             )
-        bot_id = options.id
 
+        bot_id = options.id or self.default_bot_id
         if bot_id is None:
-            await self._ensure_bot_user()
-            bot_id = self.bot_id
+            raise errors.ClientException("bot_id or default_bot_id is unset.")
 
         widget_query = f"noavatar={str(options.noavatar).lower()}"
         for key, value in options.colors.items():
@@ -388,11 +331,19 @@ class DBLClient:
         """
         if self.is_closed:
             return
-        else:
-            await self.http.close()
-            if self.autopost:
-                self.autopost_task.cancel()
-            self._is_closed = True
 
-        with suppress(AttributeError):
-            delattr(self.bot, "on_autopost_error")
+        if hasattr(self, "http"):
+            await self.http.close()
+
+        self._is_closed = True
+
+    def autopost(self) -> AutoPoster:
+        """
+        Returns a helper instance for auto-posting.
+
+        Returns
+        -------
+        autoposter: :obj:`~.autopost.AutoPoster`
+            An instance of AutoPoster.
+        """
+        return AutoPoster(self)